Create README.md
I am a graduate from the Indian Institute of Technology (IIT) Bhubaneswar, Seeking a challenging position with opportunities for career advancement to having a long career as an Analyst, gain further skills and attain the goal of organization for mutual growth.
I possess a Combination of Skills and experience that makes me stand out of the crowd. My Certifications on Power BI and Python combined with exceptional Communicational, and Organizational skills will be an added advantage for the offered role. I am capable of writing SQL queries to get insights from the data.